+++ addons/laserpointer/CfgJointRails.hpp
@@ -1,0 +1,8 @@
+class asdg_SlotInfo;
+class asdg_FrontSideRail: asdg_SlotInfo {
+    class compatibleItems {
+        ACE_acc_pointer_red = 1;
+        ACE_acc_pointer_green = 1;
+        ACE_acc_pointer_green_IR = 1;
+    };
+};
