+++ addons/optics/CfgEventHandlers.hpp
@@ -1,4 +1,3 @@
-
 class Extended_PreInit_EventHandlers {
     class ADDON {
         init = QUOTE(call COMPILE_FILE(XEH_preInit));
