+++ addons/optics/CfgJointRails.hpp
@@ -1,0 +1,15 @@
+class asdg_OpticRail;
+class asdg_OpticRail1913: asdg_OpticRail {
+    class compatibleItems {
+        ACE_optic_Hamr_2D = 1;
+        ACE_optic_Hamr_PIP = 1;
+        ACE_optic_Arco_2D = 1;
+        ACE_optic_Arco_PIP = 1;
+        ACE_optic_MRCO_2D = 1;
+        ACE_optic_MRCO_PIP = 1;
+        ACE_optic_SOS_2D = 1;
+        ACE_optic_SOS_PIP = 1;
+        ACE_optic_LRPS_2D = 1;
+        ACE_optic_LRPS_PIP = 1;
+    };
+};
