+++ addons/optics/CfgVehicles.hpp
@@ -1,4 +1,3 @@
-
 class CfgVehicles {
     class Box_NATO_Support_F;
     class ACE_Box_Misc: Box_NATO_Support_F {
