+++ addons/optics/CfgWeapons.hpp
@@ -1,18 +1,17 @@
-
 class CfgWeapons {
     class ItemCore;
     class InventoryOpticsItem_Base_F;
     class Default;
-    
+
     class Binocular: Default {
         forceOptics = 0;             // Allow using compass with Binocular
         opticsZoomMin = 0.056889;     // 5.25x power
         opticsZoomMax = 0.056889;     // 9 px/mil
-        modelOptics = "\z\ace\addons\optics\models\NWD_M22_5x"; // 7° horizontal field of view
+        modelOptics = "\z\ace\addons\optics\models\NWD_M22_5x"; // 7ï¿½ horizontal field of view
         visionMode[] = {"Normal"}; // Can't use nvgs with binoculars any more than you can with scopes
         // Fix AI using Binocs on short range - #18737
         // minRange = 300; // 300 = uses Rangefinder often (runs a few meters, stops, uses RF, repeats)
-        minRange = 500; //500 = seem almost never use it..? 
+        minRange = 500; //500 = seem almost never use it..?
         minRangeProbab = 0.001;
         midRange = 1000;
         midRangeProbab = 0.01;
