+++ addons/optics/config.cpp
@@ -30,5 +30,6 @@ class CfgPatches {
 #include "CfgRscTitles.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgWeapons.hpp"
+#include "CfgJointRails.hpp"
 
 #include "CfgPreloadTextures.hpp"
