--- optionals/compat_asdg/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\ace\addons\compat_asdg
