--- optionals/compat_asdg/script_component.hpp
@@ -1,5 +1,0 @@
-#define COMPONENT asdg_comp
-
-#include "\z\ace\addons\main\script_mod.hpp"
-
-#include "\z\ace\addons\main\script_macros.hpp"
